Reduce size of trusted codebase, provide data parsing code with a different ProtectionDomain #2
Reverting change to PlatformClassLoader, that introduced a URLClassPath, I had originally thought that modules other than java.base also had null ProtectionDomain's, however this isn't the case, this change only placed modules on the class path, which was an unintended consequence.  It appears that ProtectionDomain's are created for modules, therefore I'm reverting this change.