Update action.yml
Suggestion for improvement from Chap:

I notice that the upload-artifact action accepts an optional
with: compression-level: <0 to 9> and I think in that action.yml it is
left to the default. As all four members of this zip are already gzip-
compressed tar files, I see the zip uploader labors to squeeze out
another 0.8% compression. It would probably save cycles in your CI runs
to use compression-level: 0 at virtually no storage cost.